fix(telemetry): use float64 for scenario timestamps
Changed timestamp fields from int64 to float64 in ScenarioDetail, TelemetryDocument, and rawTelemetrySource to support fractional timestamps from Python's time.time().

Previously, Elasticsearch documents with fractional timestamps (e.g., 1735689600.123456) failed to decode and were skipped. Now these documents parse correctly.

Added test case with fractional timestamps to verify runs remain in response. Updated swagger spec to document timestamps as `number` instead of `integer`.

+++ pkg/elasticsearch/client.go
@@ -247,10 +247,10 @@ type rawTelemetrySource struct {
 	IPSecEnabled           bool              `json:"ipsec_enabled"`
 	NodeSummaryInfos       []NodeSummaryInfo `json:"node_summary_infos"`
 	Scenarios              []struct {
-		ScenarioType   string `json:"scenario_type"`
-		StartTimestamp int64  `json:"start_timestamp"`
-		EndTimestamp   int64  `json:"end_timestamp"`
-		ExitStatus     int    `json:"exit_status"`
+		ScenarioType   string  `json:"scenario_type"`
+		StartTimestamp float64 `json:"start_timestamp"`
+		EndTimestamp   float64 `json:"end_timestamp"`
+		ExitStatus     int     `json:"exit_status"`
 		// Parameters shape varies by scenario type (object keyed by scenario
 		// name, whose value may be an object or an array), so it is kept raw and
 		// searched for a namespace rather than decoded into a fixed struct.
